pyproject.toml
[tool.poetry]
name = "expensemgr"
version = "0.1.0"
description = ""
authors = ["smandalsid <smandal.sid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.120.4"}
uvicorn = {extras = ["standard"], version = "^0.38.0"}
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pytest = "^8.4.2"
factory-boy = "^3.3.3"
pytest-asyncio = "^1.2.0"
pyodbc = "^5.3.0"
alembic = "^1.17.1"
pytest-cov = "^7.0.0"
bcrypt = "4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
